requirements.txt
# requirements.txt
streamlit
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
requests
# For Streamlit's voice input
pyaudio
speechrecognition
# For the Telegram Agent
quart
python-telegram-bot
ngrok
# For Telegram's voice note transcription
google-cloud-speech
# requirements.txt
# ... (all your existing libraries) ...
toml